+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [#135](https://github.com/adamcaudill/yawast/issues/135) - Incomplete Certificate Chain Warning
 * [#137](https://github.com/adamcaudill/yawast/issues/137) - Warn on TLS 1.0
 * [#138](https://github.com/adamcaudill/yawast/issues/138) - Warn on Symantec Roots
+* [#139](https://github.com/adamcaudill/yawast/issues/139) - Add Spider Option
 * [#130](https://github.com/adamcaudill/yawast/issues/130) - Bug: HSTS Error leads to printing HTML
 * [#132](https://github.com/adamcaudill/yawast/issues/132) - Bug: Typo in SSL Output